@@ -45,7 +45,7 @@ export default function Header({sticky}) {
             <Container fluid={false}>
                 <Row>
                     <Col xs="8">
-                        <h3 id="site-title" className="animated fadeIn">ALWAYS AWAIS</h3>
+                        <h3 id="site-title" className="animated fadeIn">AWAIS WAHEED</h3>
                         <div className="designation animated flipInX" title="UX Designer - UI Designer - Font-end Developer">UX/UI DESIGNER<br/>UI DEVELOPER</div>
                     </Col>
                     <Col xs="4" className="relative">
